chore : bindingAdapter 파일 정리
