+++ app/src/main/java/com/surveasy/surveasy/presentation/bindingadapters/ButtonBindingAdapter.kt
@@ -1,16 +1,19 @@
 package com.surveasy.surveasy.presentation.bindingadapters
 
-import android.util.Log
 import androidx.appcompat.widget.AppCompatButton
 import androidx.databinding.BindingAdapter
+import com.surveasy.surveasy.presentation.intro.RegisterUiState
 
 @BindingAdapter("rewardBtn")
 fun AppCompatButton.rewardBtn(reward: Int) {
     text = "${reward}원 받으러 가기"
 }
 
-@BindingAdapter("fsNextBtn")
-fun AppCompatButton.fsNextBtn(flag : Boolean){
-    Log.d("TEST", flag.toString())
-    isEnabled = flag
+@BindingAdapter("setDoneEnable", "needEtc")
+fun AppCompatButton.setDoneEnable(uiState: RegisterUiState?, needEtc: Boolean) {
+    uiState ?: return
+    with(uiState) {
+        isEnabled =
+            inflowState.isValid && bankState.isValid && accountState.isValid && accountOwnerState.isValid && (!needEtc || inflowEtcState.isValid)
+    }
 }
